refactor: factor out makeConfig
 - prune unused takeWhile
