+++ subm/src/config.js
@@ -1,0 +1,17 @@
+/**
+ * @param { NodeJS.ProcessEnv } env
+ * @returns { TemplateTag }
+ * @typedef { (parts: TemplateStringsArray, ...args: unknown[]) => string } TemplateTag
+ */
+const makeConfig = env => {
+  return ([name], ..._args) => {
+    const value = env[name];
+    if (value === undefined) {
+      throw Error(`${name} not configured`);
+    }
+    return value;
+  };
+};
+
+/* global exports */
+exports.makeConfig = makeConfig;
